small reorganization
